refactor: extract MCP server into sapphire-journal-mcp crate
Split the MCP server out of `sapphire-journal-cli` into its own crate
`sapphire-journal-mcp`, exposing a `sapphire-journal-mcp` binary.

Crate layout:
- New `sapphire-journal-mcp` with `lib.rs` (re-exports `server::run` and
  the `shared` module) and `main.rs` (the binary entry point).
- `commands/mcp.rs` and `shared/` move out of `sapphire-journal-cli`
  into the new crate as `server.rs` and `shared/` (preserved as renames
  for git history).
- `sapphire-journal-cli` now depends on `sapphire-journal-mcp` for both
  the shared frontend helpers used by `entry.rs` and the MCP runtime.

Transitional CLI behavior:
- `sajo mcp` keeps working: the subcommand is now a thin shim that calls
  `sapphire_journal_mcp::run`. New integrations should invoke the
  `sapphire-journal-mcp` binary directly.

VSCode extension:
- Bundle `sapphire-journal-mcp` (not `sajo`) in the VSIX.
- Add `sapphire-journal.mcpBinaryPath` setting; deprecate `binaryPath`.
- Legacy `binaryPath` is still respected and invoked as `sajo mcp` so
  pre-split installs continue to work.

Release workflows:
- `release-cli.yml` publishes `sapphire-journal-mcp` between core and cli.
- `release-vscode.yml` builds and stages the new MCP binary.

Closes #201
