+++ sapphire-journal-cli/Cargo.toml
@@ -14,21 +14,30 @@ path = "src/main.rs"
 
 [features]
 default = ["sqlite-store", "lancedb-store", "fastembed-embed", "git-sync"]
-sqlite-store = ["sapphire-journal-core/sqlite-store"]
-lancedb-store = ["sapphire-journal-core/lancedb-store"]
-fastembed-embed = ["sapphire-journal-core/fastembed-embed"]
-git-sync = ["sapphire-journal-core/git-sync"]
+sqlite-store = [
+    "sapphire-journal-core/sqlite-store",
+    "sapphire-journal-mcp/sqlite-store",
+]
+lancedb-store = [
+    "sapphire-journal-core/lancedb-store",
+    "sapphire-journal-mcp/lancedb-store",
+]
+fastembed-embed = [
+    "sapphire-journal-core/fastembed-embed",
+    "sapphire-journal-mcp/fastembed-embed",
+]
+git-sync = [
+    "sapphire-journal-core/git-sync",
+    "sapphire-journal-mcp/git-sync",
+]
 
 [dependencies]
 clap.workspace = true
 sapphire-journal-core = { path = "../sapphire-journal-core", version = "0.11.1", default-features = false }
-rmcp.workspace = true
-tokio = { workspace = true, features = ["full"] }
+sapphire-journal-mcp = { path = "../sapphire-journal-mcp", version = "0.11.1", default-features = false }
+tokio = { workspace = true, features = ["rt"] }
 serde.workspace = true
 serde_json.workspace = true
-schemars.workspace = true
 chrono.workspace = true
 anyhow.workspace = true
 toml.workspace = true
-tracing.workspace = true
-tracing-subscriber.workspace = true
