+++ sapphire-journal-cli/src/main.rs
@@ -1,5 +1,4 @@
 mod commands;
-mod shared;
 
 use anyhow::Result;
 use clap::{Parser, Subcommand};
