+++ sapphire-journal-mcp/Cargo.toml
@@ -1,0 +1,37 @@
+[package]
+name = "sapphire-journal-mcp"
+edition.workspace = true
+version.workspace = true
+description = "MCP server for sapphire-journal"
+license.workspace = true
+repository.workspace = true
+categories = ["command-line-utilities"]
+keywords = ["markdown", "task-management", "notes", "mcp"]
+
+[lib]
+path = "src/lib.rs"
+
+[[bin]]
+name = "sapphire-journal-mcp"
+path = "src/main.rs"
+
+[features]
+default = ["sqlite-store", "lancedb-store", "fastembed-embed", "git-sync"]
+sqlite-store = ["sapphire-journal-core/sqlite-store"]
+lancedb-store = ["sapphire-journal-core/lancedb-store"]
+fastembed-embed = ["sapphire-journal-core/fastembed-embed"]
+git-sync = ["sapphire-journal-core/git-sync"]
+
+[dependencies]
+clap.workspace = true
+sapphire-journal-core = { path = "../sapphire-journal-core", version = "0.11.1", default-features = false }
+rmcp.workspace = true
+tokio = { workspace = true, features = ["full"] }
+serde.workspace = true
+serde_json.workspace = true
+schemars.workspace = true
+chrono.workspace = true
+anyhow.workspace = true
+toml.workspace = true
+tracing.workspace = true
+tracing-subscriber.workspace = true
