+++ sapphire-journal-mcp/src/lib.rs
@@ -1,0 +1,4 @@
+pub mod server;
+pub mod shared;
+
+pub use server::run;
