+++ sapphire-journal-mcp/src/shared/fields.rs
