+++ sapphire-journal-mcp/src/shared/filter.rs
