+++ sapphire-journal-mcp/src/shared/mod.rs
