+++ sapphire-journal-mcp/src/shared/state.rs
